Moved new Implementation to seperate folder
