+++ headers/New_Implementation/Flexible_GOL_Board.h
